requirements.txt
httpx[http2]==0.27.2
pyperclip==1.8.2
pynput==1.7.7
Jinja2==3.1.4
python-dotenv==1.0.1
tenacity==9.0.0
rich==13.8.0
tomli==2.0.1

# GUI dependencies
pillow==10.0.1